Add feature-gate for codegen
Allows RCC to be used purely for parsing, without requiring cranelift or
any other codegen backend.
