+++ Cargo.toml
@@ -15,9 +15,9 @@ documentation = "https://docs.rs/rcc"
 [dependencies]
 lazy_static = "1"
 ansi_term = { version = "0.12", optional = true }
-cranelift = "0.59"
-cranelift-module = "0.59"
-cranelift-object = "0.59"
+cranelift = { version = "0.59", optional = true }
+cranelift-module = { version = "0.59", optional = true }
+cranelift-object = { version = "0.59", optional = true }
 cranelift-simplejit = { version = "0.59", optional = true }
 env_logger = { version = "0.7", default-features = false, optional = true }
 hexponent = "0.2.2"
@@ -37,14 +37,16 @@ proptest = "0.9"
 proptest-derive = "0.1"
 
 [features]
-default = ["bin"]
-bin = ["ansi_term", "tempfile", "pico-args", "color-backtrace", "env_logger"]
-jit = ["cranelift-simplejit"]
+default = ["cc", "codegen"]
+# The `rcc` binary
+cc = ["ansi_term", "tempfile", "pico-args", "color-backtrace", "env_logger", "codegen"]
+codegen = ["cranelift", "cranelift-module", "cranelift-object"]
+jit = ["codegen", "cranelift-simplejit"]
 
 [[bin]]
 name = "rcc"
 path = "src/main.rs"
-required-features = ["bin"]
+required-features = ["cc"]
 
 [[bench]]
 name = "examples"
