+++ src/arch/mod.rs
@@ -1,21 +1,12 @@
 #![warn(missing_docs)]
 
 use std::cmp::max;
-use std::convert::TryInto;
 
-use cranelift::codegen::{
-    ir::{
-        types::{self, Type as IrType},
-        AbiParam, ArgumentPurpose, Signature,
-    },
-    isa::{CallConv, TargetIsa},
-};
-use lazy_static::lazy_static;
 use target_lexicon::Triple;
 
 use crate::data::{
     prelude::*,
-    types::{ArrayType, FunctionType, StructType},
+    types::{ArrayType, StructType},
 };
 use Type::*;
 
@@ -33,12 +24,6 @@ const CHAR_SIZE: u16 = 1;
 /// The target triple is represented as a struct and contains additional
 /// information like ABI and endianness.
 pub(crate) const TARGET: Triple = Triple::host();
-// TODO: make this const when const_if_match is stabilized
-// TODO: see https://github.com/rust-lang/rust/issues/49146
-lazy_static! {
-    /// The calling convention for the current target.
-    pub(crate) static ref CALLING_CONVENTION: CallConv = CallConv::triple_default(&TARGET);
-}
 
 mod x64;
 pub(crate) use x64::*;
@@ -174,78 +159,6 @@ impl Type {
             Void => Err("cannot take `alignof` void"),
             VaList => Err("cannot take `alignof` va_list"),
             Error => Err("cannot take `alignof` <type error>"),
-        }
-    }
-    /// Return an IR integer type large enough to contain a pointer.
-    pub fn ptr_type() -> IrType {
-        IrType::int(CHAR_BIT * PTR_SIZE).expect("pointer size should be valid")
-    }
-    /// Return an IR type which can represent this C type
-    pub fn as_ir_type(&self) -> IrType {
-        match self {
-            // Integers
-            Bool => types::B1,
-            Char(_) | Short(_) | Int(_) | Long(_) | Pointer(_) | Enum(_, _) => {
-                let int_size = SIZE_T::from(CHAR_BIT)
-                    * self
-                        .sizeof()
-                        .expect("integers should always have a valid size");
-                IrType::int(int_size.try_into().unwrap_or_else(|_| {
-                    panic!(
-                        "integers should never have a size larger than {}",
-                        i16::max_value()
-                    )
-                }))
-                .unwrap_or_else(|| panic!("unsupported size for IR: {}", int_size))
-            }
-
-            // Floats
-            // TODO: this is hard-coded for x64
-            Float => types::F32,
-            Double => types::F64,
-
-            // Aggregates
-            // arrays and functions decay to pointers
-            Function(_) | Array(_, _) => IrType::int(PTR_SIZE * CHAR_BIT)
-                .unwrap_or_else(|| panic!("unsupported size of IR: {}", PTR_SIZE)),
-            // void cannot be loaded or stored
-            _ => types::INVALID,
-        }
-    }
-}
-
-impl FunctionType {
-    /// Generate the IR function signature for `self`
-    pub fn signature(&self, isa: &dyn TargetIsa) -> Signature {
-        let mut params = if self.params.len() == 1 && self.params[0].ctype == Type::Void {
-            // no arguments
-            Vec::new()
-        } else {
-            self.params
-                .iter()
-                .map(|param| AbiParam::new(param.ctype.as_ir_type()))
-                .collect()
-        };
-        if self.varargs {
-            let al = isa
-                .register_info()
-                .parse_regunit("rax")
-                .expect("x86 should have an rax register");
-            params.push(AbiParam::special_reg(
-                types::I8,
-                ArgumentPurpose::Normal,
-                al,
-            ));
-        }
-        let return_type = if !self.should_return() {
-            vec![]
-        } else {
-            vec![AbiParam::new(self.return_type.as_ir_type())]
-        };
-        Signature {
-            call_conv: *CALLING_CONVENTION,
-            params,
-            returns: return_type,
         }
     }
 }
