+++ src/ir/mod.rs
@@ -5,20 +5,71 @@ mod stmt;
 use std::collections::{HashMap, VecDeque};
 use std::convert::TryFrom;
 
-use crate::data::{prelude::*, types::FunctionType, Initializer, Scope, StorageClass};
+use crate::arch::{CHAR_BIT, PTR_SIZE, SIZE_T, TARGET};
+use crate::data::{
+    lex::ComparisonToken, prelude::*, types::FunctionType, Initializer, Scope, StorageClass,
+};
 use cranelift::codegen::{
     self,
     ir::{
+        condcodes::{FloatCC, IntCC},
         entities::StackSlot,
         function::Function,
         stackslot::{StackSlotData, StackSlotKind},
-        ExternalName, InstBuilder, MemFlags,
+        types::{self, Type as IrType},
+        AbiParam, ArgumentPurpose, ExternalName, InstBuilder, MemFlags, Signature,
     },
-    settings,
+    isa::{CallConv, TargetIsa},
+    settings::{self, Configurable, Flags},
 };
 use cranelift::frontend::Switch;
-use cranelift::prelude::{Block, FunctionBuilder, FunctionBuilderContext, Signature};
+use cranelift::prelude::{Block, FunctionBuilder, FunctionBuilderContext};
 use cranelift_module::{self, Backend, DataId, FuncId, Linkage, Module};
+use cranelift_object::{ObjectBackend, ObjectBuilder, ObjectTrapCollection};
+use lazy_static::lazy_static;
+
+// TODO: make this const when const_if_match is stabilized
+// TODO: see https://github.com/rust-lang/rust/issues/49146
+lazy_static! {
+    /// The calling convention for the current target.
+    pub(crate) static ref CALLING_CONVENTION: CallConv = CallConv::triple_default(&TARGET);
+}
+
+pub(crate) fn get_isa(jit: bool) -> Box<dyn TargetIsa + 'static> {
+    let mut flags_builder = cranelift::codegen::settings::builder();
+    // `simplejit` requires non-PIC code
+    if !jit {
+        // allow creating shared libraries
+        flags_builder
+            .enable("is_pic")
+            .expect("is_pic should be a valid option");
+    }
+    // use debug assertions
+    flags_builder
+        .enable("enable_verifier")
+        .expect("enable_verifier should be a valid option");
+    // minimal optimizations
+    flags_builder
+        .set("opt_level", "speed")
+        .expect("opt_level: speed should be a valid option");
+    // don't emit call to __cranelift_probestack
+    flags_builder
+        .set("enable_probestack", "false")
+        .expect("enable_probestack should be a valid option");
+    let flags = Flags::new(flags_builder);
+    cranelift::codegen::isa::lookup(TARGET)
+        .unwrap_or_else(|_| panic!("platform not supported: {}", TARGET))
+        .finish(flags)
+}
+
+pub fn initialize_aot_module(name: String) -> Module<ObjectBackend> {
+    Module::new(ObjectBuilder::new(
+        get_isa(false),
+        name,
+        ObjectTrapCollection::Disabled,
+        cranelift_module::default_libcall_names(),
+    ))
+}
 
 enum Id {
     Function(FuncId),
@@ -329,5 +380,110 @@ impl<B: Backend> Compiler<B> {
 impl FunctionType {
     fn has_params(&self) -> bool {
         !(self.params.len() == 1 && self.params[0].ctype == Type::Void)
+    }
+
+    /// Generate the IR function signature for `self`
+    pub fn signature(&self, isa: &dyn TargetIsa) -> Signature {
+        let mut params = if self.params.len() == 1 && self.params[0].ctype == Type::Void {
+            // no arguments
+            Vec::new()
+        } else {
+            self.params
+                .iter()
+                .map(|param| AbiParam::new(param.ctype.as_ir_type()))
+                .collect()
+        };
+        if self.varargs {
+            let al = isa
+                .register_info()
+                .parse_regunit("rax")
+                .expect("x86 should have an rax register");
+            params.push(AbiParam::special_reg(
+                types::I8,
+                ArgumentPurpose::Normal,
+                al,
+            ));
+        }
+        let return_type = if !self.should_return() {
+            vec![]
+        } else {
+            vec![AbiParam::new(self.return_type.as_ir_type())]
+        };
+        Signature {
+            call_conv: *CALLING_CONVENTION,
+            params,
+            returns: return_type,
+        }
+    }
+}
+
+impl ComparisonToken {
+    pub fn to_int_compare(self, signed: bool) -> IntCC {
+        use ComparisonToken::*;
+        match (self, signed) {
+            (Less, true) => IntCC::SignedLessThan,
+            (Less, false) => IntCC::UnsignedLessThan,
+            (LessEqual, true) => IntCC::SignedLessThanOrEqual,
+            (LessEqual, false) => IntCC::UnsignedLessThanOrEqual,
+            (Greater, true) => IntCC::SignedGreaterThan,
+            (Greater, false) => IntCC::UnsignedGreaterThan,
+            (GreaterEqual, true) => IntCC::SignedGreaterThanOrEqual,
+            (GreaterEqual, false) => IntCC::UnsignedGreaterThanOrEqual,
+            (EqualEqual, _) => IntCC::Equal,
+            (NotEqual, _) => IntCC::NotEqual,
+        }
+    }
+    pub fn to_float_compare(self) -> FloatCC {
+        use ComparisonToken::*;
+        match self {
+            Less => FloatCC::LessThan,
+            LessEqual => FloatCC::LessThanOrEqual,
+            Greater => FloatCC::GreaterThan,
+            GreaterEqual => FloatCC::GreaterThanOrEqual,
+            EqualEqual => FloatCC::Equal,
+            NotEqual => FloatCC::NotEqual,
+        }
+    }
+}
+
+use std::convert::TryInto;
+impl Type {
+    /// Return an IR integer type large enough to contain a pointer.
+    pub fn ptr_type() -> IrType {
+        IrType::int(CHAR_BIT * PTR_SIZE).expect("pointer size should be valid")
+    }
+    /// Return an IR type which can represent this C type
+    pub fn as_ir_type(&self) -> IrType {
+        use Type::*;
+
+        match self {
+            // Integers
+            Bool => types::B1,
+            Char(_) | Short(_) | Int(_) | Long(_) | Pointer(_) | Enum(_, _) => {
+                let int_size = SIZE_T::from(CHAR_BIT)
+                    * self
+                        .sizeof()
+                        .expect("integers should always have a valid size");
+                IrType::int(int_size.try_into().unwrap_or_else(|_| {
+                    panic!(
+                        "integers should never have a size larger than {}",
+                        i16::max_value()
+                    )
+                }))
+                .unwrap_or_else(|| panic!("unsupported size for IR: {}", int_size))
+            }
+
+            // Floats
+            // TODO: this is hard-coded for x64
+            Float => types::F32,
+            Double => types::F64,
+
+            // Aggregates
+            // arrays and functions decay to pointers
+            Function(_) | Array(_, _) => IrType::int(PTR_SIZE * CHAR_BIT)
+                .unwrap_or_else(|| panic!("unsupported size of IR: {}", PTR_SIZE)),
+            // void cannot be loaded or stored
+            _ => types::INVALID,
+        }
     }
 }
